Copy module sources to builddir and don't set src= for make command
Now we set the `M` variable with the path to the empty and writable
`builddir` directory where all build files (*.o, *.ko, ...) are created,
and the `src` variable with the path to the readonly `srcdir` with the
module source files.

Unfortunately, this doesn't work anymore due to the following changes in
the recent kernel make files:

- https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=9a0ebe5011f49e932bb0a2cea2034fd65e6e567e
- https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=b1992c3772e69a6fd0e3fc81cd4d2820c8b6eca0

So we have to copy the module sources from `srcdir` to `builddir`.

See: https://gitlab.alpinelinux.org/alpine/aports/-/issues/16718

Co-Authored-By: Naomi Rennie-Waldock <naomi.renniewaldock@gmail.com>